% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{doi_check}
\alias{doi_check}
\title{Check if a string is a DOI}
\usage{
doi_check(doi_vec, na_yields_false = FALSE)
}
\arguments{
\item{doi_vec}{A character vector containing DOIs.}

\item{na_yields_false}{A single logical value stating how to deal with NAs in the input vector.
Set to FALSE, doi_check will return NA. Set to TRUE, it will return FALSE.}
}
\value{
A logical vector with TRUE indicating whether an element of the input vector is a valid DOI.
}
\description{
Returns TRUE if a given string is a valid DOI. Vectorised over doi_vec.
}
