% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{doi_detect}
\alias{doi_detect}
\title{Detect the presence or absence of a DOI in a string}
\usage{
doi_detect(doi_vec, na_yields_false = FALSE)
}
\arguments{
\item{doi_vec}{A character vector containing DOIs.}

\item{na_yields_false}{A single logical value stating how to deal with NAs in the input vector.
Set to FALSE, doi_detect will return NA. Set to TRUE, it will return FALSE.}
}
\value{
A logical vector with TRUE indicating whether a DOI was found.
}
\description{
Returns TRUE if a given string contains a valid DOI. Vectorised over doi_vec.
}
