% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{doi_extract_all}
\alias{doi_extract_all}
\title{Extracts all DOIs from a string}
\usage{
doi_extract_all(doi_vec, lower_case = FALSE, simplify = FALSE)
}
\arguments{
\item{doi_vec}{A character vector.}

\item{lower_case}{A single logical value stating whether or not to put the whole DOI in lower case.
As DOIs are case insensitive, using lower case for all may help to identify duplicates more easily.}

\item{simplify}{A single logical value stating whether or not to flatten the resulting list into
a character vector with unique DOIs only. This is useful if you just want to retrieve all unique
DOIs from a text and you do not mind loosing the mapping from the input vector to the output.}
}
\value{
A list of character vectors containing DOIs (if found) or NAs otherwise. The length of the
list equals the length of the input vector. In case simplify = TRUE, a single character vector is
returned, whose length is not related to the length of the input vector.
}
\description{
Extracts all valid DOIs from a given character vector. Vectorised over doi_vec.
}
