% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{doi_extract_first}
\alias{doi_extract_first}
\title{Extracts the first DOI from a string}
\usage{
doi_extract_first(doi_vec, lower_case = FALSE)
}
\arguments{
\item{doi_vec}{A character vector.}

\item{lower_case}{A single logical value stating whether or not to put the whole DOI in lower case.
As DOIs are case insensitive, using lower case for all may help to identify duplicates more easily.}
}
\value{
A character vector containing DOIs (if found) or NAs otherwise.
}
\description{
Extracts valid DOIs from a given character vector. Vectorised over doi_vec.
Only returns the first match for each element.
}
