% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altmetric.R
\name{doi_get_altmetrics}
\alias{doi_get_altmetrics}
\alias{doi_get_altmetrics_single}
\alias{doi_get_altmetrics_background}
\title{Fetch publication data from the Altmetric API}
\usage{
doi_get_altmetrics(doi_vec, api_key = NULL, wait_for = 1)

doi_get_altmetrics_single(doi_vec, api_key = NULL)

doi_get_altmetrics_background(doi_vec, api_key = NULL)
}
\arguments{
\item{doi_vec}{A character vector containing publication DOIs. doi_get_altmetrics_single only uses the
first element.}

\item{api_key}{The Altmetric api key to be used for the query.
If no api key is provided, the access is currently limited to 720 calls per hour and 1200 calls per day.}

\item{wait_for}{Time to wait between each api call in seconds. The default value is 1 which is what Altmetric
urges you not to undercut. doi_get_altmetrics_background uses this as a fixed value.}
}
\value{
A list of two tibbles with the same number of rows. The first, called metrics, contains the downloaded
publication meta data for the supplied DOIs. The second one, called api_info, provides data on the status of each
API call. It has four columns:
\itemize{
\item \code{api_response_code} The response code of the API
\item \code{api_response_msg} The description of the response code, which can either be an error message or the word 'Success'
\item \code{remaining_calls_this_hour} The number of remaining API calls beeing until the hourly rate limit is reached
\item \code{remaining_calls_today} The number of remaining API calls beeing until the daily rate limit is reached
}
}
\description{
The Altmetric API provides a set of publication meta data including attention scores in social networks
for a large number of publications. See \url{https://api.altmetric.com} for more
information. Use these functions to retrieve Altmetric data by supplying the DOI (digital object identifier)
of a publication.
}
\details{
\itemize{
\item \code{doi_get_altmetrics} is vectorized over doi_vec and allows downloading metrics of multiple publications. It
provides a simple text progress bar.
\item \code{doi_get_altmetrics_single} fetches metrics for a single publication only. It is suited for non-interactive
sessions like a shiny app that implements a custom progress indicator.
\item \code{doi_get_altmetrics_background} is suited for background jobs. It allows downloading a large number of
publications that usually would exceed the API's rate limits. If the rate limit is reached, it waits
until retrieval is possible again and then continues downloading until completion. Intermediate results will be
saved in a file and can be backed up if the execution is interrupted.
}
}
\examples{
my_dois <- c("10.7191/jeslib.2021.1180", "10.1108/02640470610689151", ",,,not,a,doi,,,")

result <- doi_get_altmetrics(my_dois)
result

# successful calls only
result$metrics[result$api_info$api_response_msg == "Success", ]

# failed calls only
result$metrics[result$api_info$api_response_msg != "Success", ]

# mean Altmetric score of all DOIs
mean(result$metrics$score, na.rm = TRUE)
}
