% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{doi_set_regex}
\alias{doi_set_regex}
\title{Set the regular expression used to find a DOI}
\usage{
doi_set_regex(rg_expr = "10\\\\.\\\\d{4,9}/[-._;()/:a-z0-9A-Z]+")
}
\arguments{
\item{rg_expr}{A single string containing the regular expression to use further on during a session.}
}
\description{
The doi package uses a regular expression to find or extract DOIs.
Its standard value is "10\\\\.\\\\d{4,9}/[-._;()/:a-z0-9A-Z]+".
Use this function to set a custom regular expression of your choice, valid for the current session.
Omit the argument to set the regular expression back to standard.
}
