% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{doi_to_url}
\alias{doi_to_url}
\title{Turn DOIs into URLs}
\usage{
doi_to_url(doi_vec)
}
\arguments{
\item{doi_vec}{A character vector containing DOIs.}
}
\value{
A character vector containing URLs.
}
\description{
Turns DOIs into URLs by placing the prefix "https://www.doi.org/" in front of them.
}
